requirements.txt
azure-ai-textanalytics
